requirements.txt
# Runtime (what you already have)
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
drf-yasg>=1.21.1,<2.0
django-environ>=0.9.0,<1.0
psycopg2-binary>=2.9.5,<3.0
Faker>=18.3.0,<19.0
django-filter>=23.2,<24.0
djangorestframework-simplejwt>=5.2.2,<6.0
whitenoise>=6.0,<7.0
gunicorn>=20.1.0,<21.0

# Optional / Helpful middleware
django-cors-headers>=3.13.0,<4.0       # if you need Cross-Origin support
django-debug-toolbar>=4.0,<5.0        # profiling & debug panels in development

# Dev & Testing
django-extensions>=3.2,<4.0           # shell_plus, graph_models, etc.
pytest>=7.0,<8.0
pytest-django>=4.0,<5.0
coverage>=6.0,<7.0
factory-boy>=3.2,<4.0                 # model factories for tests

# Linting / Formatting (optional)
flake8>=6.0,<7.0
black>=24.0,<25.0
isort>=6.0,<7.0
